tambah : testing oke

+++ database/migrations/2025_08_25_180117_add_is_logged_in_to_users_table.php
@@ -10,17 +10,21 @@ return new class extends Migration
      * Run the migrations.
      */
     public function up()
-{
-    Schema::table('users', function (Blueprint $table) {
-        $table->boolean('is_logged_in')->default(false);
-    });
-}
+    {
+        Schema::table('users', function (Blueprint $table) {
+            if (!Schema::hasColumn('users', 'is_logged_in')) {
+                $table->boolean('is_logged_in')->default(false);
+            }
+        });
+    }
 
-public function down()
-{
-    Schema::table('users', function (Blueprint $table) {
-        $table->dropColumn('is_logged_in');
-    });
-}
+    public function down()
+    {
+        Schema::table('users', function (Blueprint $table) {
+            if (Schema::hasColumn('users', 'is_logged_in')) {
+                $table->dropColumn('is_logged_in');
+            }
+        });
+    }
 
 };
