+++ database/migrations/2025_08_25_180334_add_logged_in_at_to_users_table.php
@@ -7,17 +7,21 @@ use Illuminate\Support\Facades\Schema;
 return new class extends Migration
 {
     public function up()
-{
-    Schema::table('users', function (Blueprint $table) {
-        $table->timestamp('logged_in_at')->nullable();
-    });
-}
+    {
+        Schema::table('users', function (Blueprint $table) {
+            if (!Schema::hasColumn('users', 'logged_in_at')) {
+                $table->timestamp('logged_in_at')->nullable();
+            }
+        });
+    }
 
-public function down()
-{
-    Schema::table('users', function (Blueprint $table) {
-        $table->dropColumn('logged_in_at');
-    });
-}
+    public function down()
+    {
+        Schema::table('users', function (Blueprint $table) {
+            if (Schema::hasColumn('users', 'logged_in_at')) {
+                $table->dropColumn('logged_in_at');
+            }
+        });
+    }
 
 };
